# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "eBay Sales Tool API",
    "version": "1.0.0",
    "description": "API documentation for the eBay Sales Tool"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/api/customers": {
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Duplicate email"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "summary": "Get all customers (with optional filters and pagination)",
        "tags": [
          "Customers"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by customer status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Page size for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "customers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/catalog": {
      "post": {
        "summary": "Create a new catalog entry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the item"
                  },
                  "price": {
                    "type": "number",
                    "description": "The price of the item"
                  },
                  "description": {
                    "type": "string",
                    "description": "A description of the item"
                  },
                  "manufacturer_info": {
                    "type": "string",
                    "description": "Additional information about the manufacturer"
                  },
                  "manufacturer": {
                    "type": "string",
                    "description": "The manufacturer of the item"
                  },
                  "model": {
                    "type": "string",
                    "description": "The model of the item"
                  },
                  "serial_number": {
                    "type": "string",
                    "description": "The serial number of the item"
                  },
                  "product_page_link": {
                    "type": "string",
                    "description": "A link to the product page"
                  },
                  "dimension": {
                    "type": "object",
                    "properties": {
                      "x": {
                        "type": "number",
                        "description": "The length of the item"
                      },
                      "y": {
                        "type": "number",
                        "description": "The width of the item"
                      },
                      "z": {
                        "type": "number",
                        "description": "The height of the item"
                      }
                    }
                  },
                  "weight": {
                    "type": "number",
                    "description": "The weight of the item"
                  },
                  "condition": {
                    "type": "string",
                    "description": "The condition of the item"
                  },
                  "category": {
                    "type": "string",
                    "description": "The category of the item"
                  },
                  "sku_barcode": {
                    "type": "string",
                    "description": "The SKU or barcode of the item"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "A list of image URLs for the item"
                  },
                  "specifications": {
                    "type": "string",
                    "description": "Specifications of the item"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Catalog entry created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "The ID of the created catalog entry"
                    },
                    "name": {
                      "type": "string",
                      "description": "The name of the item"
                    },
                    "price": {
                      "type": "number",
                      "description": "The price of the item"
                    },
                    "description": {
                      "type": "string",
                      "description": "A description of the item"
                    },
                    "manufacturer": {
                      "type": "string",
                      "description": "The manufacturer of the item"
                    },
                    "model": {
                      "type": "string",
                      "description": "The model of the item"
                    },
                    "serial_number": {
                      "type": "string",
                      "description": "The serial number of the item"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error creating catalog entry"
          }
        }
      }
    },
    "/api/listings": {
      "get": {
        "summary": "Retrieve all listings",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by listing status"
          },
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by item ID"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Page size for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of all listings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "properties": {
                    "listings": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Listing"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  },
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "The listing ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "The title of the listing"
                      },
                      "price": {
                        "type": "number",
                        "description": "The price of the listing"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving listings"
          }
        }
      },
      "post": {
        "summary": "Create a new listing",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Listing"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Error creating listing"
          }
        }
      }
    },
...
    "/api/listings/{id}/details": {
      "get": {
        "summary": "Get comprehensive listing details including change history",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "integer" }, "description": "Listing ID" },
          { "in": "query", "name": "history_limit", "schema": { "type": "integer" }, "description": "Optional requested number of history rows (capped by config history_display_limit)" },
          { "in": "query", "name": "history_offset", "schema": { "type": "integer" }, "description": "Optional offset for history pagination (default 0)" }
        ],
        "responses": {
          "200": {
            "description": "Listing details payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "listing": { "$ref": "#/components/schemas/Listing" },
                    "catalog": { "$ref": "#/components/schemas/Catalog" },
                    "sales": { "type": "array", "items": { "$ref": "#/components/schemas/Sale" } },
                    "ownerships": { "type": "array", "items": { "$ref": "#/components/schemas/Ownership" } },
                    "change_history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "action": { "type": "string", "enum": ["create","update","delete"] },
                          "changed_fields": { "type": "array", "items": { "type": "string" } },
                          "created_at": { "type": "string", "format": "date-time" },
                          "before_data": { "type": "object" },
                          "after_data": { "type": "object" }
                        }
                      }
                    },
                    "change_history_total": { "type": "integer" },
                    "change_history_limit": { "type": "integer", "description": "Configured max history_display_limit" },
                    "change_history_requested_limit": { "type": "integer", "nullable": true },
                    "change_history_effective_limit": { "type": "integer" },
                    "change_history_offset": { "type": "integer" }
                  }
                }
              }
            }
          },
          "404": { "description": "Listing not found" },
          "500": { "description": "Error fetching listing details" }
        }
      }
    },
    "/api/ownership": {
      "post": {
        "summary": "Create a new ownership record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownershipType": {
                    "type": "string",
                    "enum": [
                      "Self",
                      "Company"
                    ],
                    "description": "The type of ownership"
                  },
                  "contact": {
                    "type": "object",
                    "properties": {
                      "firstName": {
                        "type": "string"
                      },
                      "lastName": {
                        "type": "string"
                      },
                      "address": {
                        "type": "string"
                      },
                      "telephone": {
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      }
                    }
                  },
                  "companyDetails": {
                    "type": "object",
                    "properties": {
                      "companyName": {
                        "type": "string"
                      },
                      "companyAddress": {
                        "type": "string"
                      },
                      "companyTelephone": {
                        "type": "string"
                      },
                      "companyEmail": {
                        "type": "string"
                      },
                      "assignedContact": {
                        "type": "object",
                        "properties": {
                          "firstName": {
                            "type": "string"
                          },
                          "lastName": {
                            "type": "string"
                          },
                          "telephone": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ownership record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "The ID of the created ownership record"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error creating ownership record"
          }
        }
      },
      "get": {
        "summary": "Get all ownerships/agreements (with optional filters and pagination)",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by status"
          },
          {
            "in": "query",
            "name": "agreementType",
            "schema": {
              "type": "string"
            },
            "description": "Filter by agreement type"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Page size for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of ownerships/agreements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ownerships": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Ownership"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "page": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching ownerships/agreements"
          }
        }
      }
    },
...
    "/api/sales": {
      "post": {
        "summary": "Create a new sale",
        "tags": [
          "Sales"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "soldDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "owner": {
                    "type": "string"
                  },
                  "negotiatedTerms": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sale created successfully"
          },
          "500": {
            "description": "Error creating sale"
          }
        }
      },
      "get": {
        "summary": "Get all sales (with optional filters)",
        "tags": [
          "Sales"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by sale status"
          },
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by item ID"
          },
          {
            "in": "query",
            "name": "customerId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by customer ID"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by start date"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Filter by end date"
          }
        ],
        "responses": {
          "200": {
            "description": "List of sales"
          },
          "500": {
            "description": "Error fetching sales"
          }
        }
      }
    },
...
  },
...
}